Removed insanely complex structure
sorry. Im not big brain enough
