+++ OhMyML-Project-Solution/OhMyML/Examples/OhMyML_Runnable_Example-MultipleLinearRegression.cs
@@ -14,24 +14,24 @@ namespace OhMyML
 		static void main()
 		{
 			// Example of data input
-			double[,] X = { { 1, 2, 3},
+			double[,] x = { { 1, 2, 3},
 				{ 2, 9, 11 },
 				{ 56, 111, 66}};
 
 			double[,] y = { { 6 }, { 6 }, { 11 } };
 
-			// Creating a new Supervised Learning Object with the type LinearRegressor
-			SupervisedLearning<double[,], MultipleLinearRegressor, double> sl = new SupervisedLearning<double[,], MultipleLinearRegressor, double>();
+			// Creating a new MultipleLinearRegressor
+			MultipleLinearRegressor mlr = new MultipleLinearRegressor();
 
-			// Setting the input by creating a holder of the data
-			sl.SetInput(OhMath.CreateHolder(X.ToEnumerable<double[,]>(), y.ToEnumerable<double[,]>()));
+			// Fitting the x and y
+			mlr.Fit(x, y);
 
 			// Getting the output predictions from our learning container
 			double[,] sample = {{3}, {5}, {7}};
-			// double prediction = sl.GetOutput(sample.ToEnumerable<double[,]>());
+			double prediction = mlr.Predict(sample);
 
 			// Displaying the output in console
-			// Console.WriteLine($"Prediction: {prediction}");
+			Console.WriteLine($"Prediction: {prediction}");
 		}
 	}
 }
