+++ OhMyML-Project-Solution/OhMyML/SourceCode/MLType/Supervised Learning/Regressor.cs
@@ -6,15 +6,15 @@ using System.Threading.Tasks;
 
 namespace OhMyML.SourceCode.MLType.Supervised_Learning
 {
-    public interface IRegressor{}
-    public interface Regressor<I> : IRegressor
-    {
-        public void Fit(I x, I y);
-        public I Predict(I x);
-    }
-    public interface Regressor<I, O> : IRegressor
-    {
-        public void Fit(I x, I y);
-        public O Predict(I x);
-    }
+	public interface IRegressor { }
+	public interface Regressor<I> : IRegressor
+	{
+		public void Fit(I x, I y);
+		public I Predict(I x);
+	}
+	public interface Regressor<I, O> : IRegressor
+	{
+		public void Fit(I x, I y);
+		public O Predict(I x);
+	}
 }
